:recycle: common selects to avoid circular references
